+++ packages/video-segment-generator/src/utils/segment.js
@@ -3,21 +3,21 @@ import { Payload } from '@tvkitchen/base-classes'
 import { INTERVALS } from '../constants'
 
 /**
- * Calculates the position associated with the origin time when compared to a reference
- * position / timestamp pair.  This origin position might be negative (e.g. if the origin
- * happened before the start of the reference stream)
+ * Calculates the interal segmenter position associated with a given frame's zero reference.
+ * This starting position might be negative (e.g. if the zero reference happened before
+ * the start of the reference stream).
  *
- * @param  {String} originTimestamp    The ISO string associated with the origin time.
- * @param  {String} referenceTimestamp The ISO string associated with the reference time.
- * @param  {Number} referencePosition  The payload associated with th ereference time.
- * @return {Number}                    The origin payload when compared to the reference stream.
+ * @param  {String} startingTimestamp      The ISO string associated with the segmenter internal
+ *                                         zero position.
+ * @param  {String} zeroReferenceTimestamp The ISO string associated with an external zero position.
+ * @return {Number}                        The position associated with the segmenter's start.
  */
-export const calculateOriginPosition = (originTimestamp, referenceTimestamp, referencePosition) => (
-	dayjs(originTimestamp).diff(dayjs(referenceTimestamp)) + referencePosition
+export const calculateStartingPosition = (startingTimestamp, zeroReferenceTimestamp) => (
+	dayjs(startingTimestamp).diff(dayjs(zeroReferenceTimestamp))
 )
 
 /**
- * Generates a complete set of payments for the period starting at a given position.
+ * Generates a complete set of payments starting from a given position.
  *
  * Segments can either be represented as an integer (the offset) or as an object with offset
  * and data values.
@@ -26,7 +26,7 @@ export const calculateOriginPosition = (originTimestamp, referenceTimestamp, ref
  * @param  {object[]} segments     The segments being used to generate payloads.
  * @return {Payload[]}             The resulting segment payloads.
  */
-export const generatePeriodSegmentPayloads = (periodPosition, segments) => segments.map(
+export const generateSegmentPayloadsFromPosition = (periodPosition, segments) => segments.map(
 	(segment) => {
 		const data = Number.isInteger(segment) ? null : Buffer.from(JSON.stringify(segment.data))
 		const offset = Number.isInteger(segment) ? segment : segment.offset
@@ -44,16 +44,19 @@ export const generatePeriodSegmentPayloads = (periodPosition, segments) => segme
  *
  * A zero value for interval will always return the originPosition.
  *
- * @param  {Number} position       The position being checked
- * @param  {Number} originPosition The starting position of the first period
- * @param  {Number} interval       The period interval size
- * @return {Number}                The starting position of the containing period
+ * @param  {Number} position                  The position being checked
+ * @param  {Number} segmenterStartingPosition The starting position of the first period
+ * @param  {Number} interval                  The period interval size
+ * @return {Number}                           The starting position of the containing period
  */
-export const getPeriodPosition = (position, originPosition, interval) => (
-	(interval === INTERVALS.INFINITE)
-		? originPosition
-		: Math.floor((position - originPosition) / interval) * interval + originPosition
-)
+export const getPeriodPosition = (position, segmenterStartingPosition, interval) => {
+	if (interval === INTERVALS.INFINITE) {
+		return segmenterStartingPosition
+	}
+	const offsetPosition = Math.floor((position - segmenterStartingPosition) / interval) * interval
+	return segmenterStartingPosition + offsetPosition
+}
+
 
 /**
  * Creates an array of values.
@@ -78,34 +81,45 @@ const range = (start, count, step) => Array.from(
  * The start and end position are inclusive, which means that segments whose positions match
  * the start or end position will be included in the result.
  *
- * @param  {Number} startPosition  The earliest segment position to generate.
- * @param  {Number} endPosition    The latest segment position to generate.
- * @param  {Number} originPosition The origin position, used to calculate segment positions.
- * @param  {Number} interval       The interval being used to generate segment sets.
- * @param  {object[]} segments     The segments being generated.
- * @return {Payload[]}             The resulting Payloads
+ * @param  {Number} rangeStartPosition        The earliest segment position to generate.
+ * @param  {Number} rangeEndPosition          The latest segment position to generate.
+ * @param  {Number} segmenterStartingPosition The zero position of the segmenter.
+ * @param  {Number} interval                  The interval being used to generate segment sets.
+ * @param  {object[]} segments                The segments being generated.
+ * @return {Payload[]}                        The resulting Payloads
  */
-export const generatePositionRangeSegmentPayloads = (
-	startPosition,
-	endPosition,
-	originPosition,
+export const generateSegmentPayloadsForPositionRange = (
+	rangeStartPosition,
+	rangeEndPosition,
+	segmenterStartingPosition,
 	interval,
 	segments,
 ) => {
 	let segmentPayloads = []
 	if (interval === INTERVALS.INFINITE) {
-		segmentPayloads = generatePeriodSegmentPayloads(originPosition, segments)
+		segmentPayloads = generateSegmentPayloadsFromPosition(
+			segmenterStartingPosition,
+			segments,
+		)
 	} else {
-		const startPeriodPosition = getPeriodPosition(startPosition, originPosition, interval)
-		const endPeriodPosition = getPeriodPosition(endPosition, originPosition, interval)
+		const startPeriodPosition = getPeriodPosition(
+			rangeStartPosition,
+			segmenterStartingPosition,
+			interval,
+		)
+		const endPeriodPosition = getPeriodPosition(
+			rangeEndPosition,
+			segmenterStartingPosition,
+			interval,
+		)
 		const periodCount = Math.floor((endPeriodPosition - startPeriodPosition) / interval) + 1
 		const periodPositions = range(startPeriodPosition, periodCount, interval)
 		segmentPayloads = periodPositions.flatMap(
-			(periodPosition) => generatePeriodSegmentPayloads(periodPosition, segments),
+			(periodPosition) => generateSegmentPayloadsFromPosition(periodPosition, segments),
 		)
 	}
 	return segmentPayloads.filter((segment) => (
-		segment.position >= startPosition
-		&& segment.position <= endPosition
+		segment.position >= rangeStartPosition
+		&& segment.position <= rangeEndPosition
 	))
 }